use two different components for slideshow template
