+++ app/templates/exhibit.hbs
@@ -1,1 +1,1 @@
-{{photo-slide exhibit=model}}
+{{slide-selector exhibit=model}}
